sensor: tdk: Add support for icm42686
Support both icm42686 and icm42688 through the same driver
