+++ drivers/sensor/CMakeLists.txt
@@ -41,7 +41,7 @@ add_subdirectory_ifdef(CONFIG_HP206C hp206c)
 add_subdirectory_ifdef(CONFIG_HS300X hs300x)
 add_subdirectory_ifdef(CONFIG_ICM42605 tdk/icm42605)
 add_subdirectory_ifdef(CONFIG_ICM42670 tdk/icm42670)
-add_subdirectory_ifdef(CONFIG_ICM42688 tdk/icm42688)
+add_subdirectory_ifdef(CONFIG_ICM4268X tdk/icm4268x)
 add_subdirectory_ifdef(CONFIG_ICP10125 tdk/icp10125)
 add_subdirectory_ifdef(CONFIG_IST8310 ist8310)
 add_subdirectory_ifdef(CONFIG_ISL29035 isl29035)
