+++ drivers/sensor/Kconfig
@@ -125,7 +125,7 @@ source "drivers/sensor/hp206c/Kconfig"
 source "drivers/sensor/hs300x/Kconfig"
 source "drivers/sensor/tdk/icm42605/Kconfig"
 source "drivers/sensor/tdk/icm42670/Kconfig"
-source "drivers/sensor/tdk/icm42688/Kconfig"
+source "drivers/sensor/tdk/icm4268x/Kconfig"
 source "drivers/sensor/tdk/icp10125/Kconfig"
 source "drivers/sensor/isl29035/Kconfig"
 source "drivers/sensor/ist8310/Kconfig"
