+++ drivers/sensor/tdk/icm4268x/icm4268x_rtio.c
@@ -7,22 +7,22 @@
 
 #include <zephyr/drivers/sensor.h>
 #include <zephyr/rtio/work.h>
-#include "icm42688.h"
-#include "icm42688_decoder.h"
-#include "icm42688_reg.h"
-#include "icm42688_rtio.h"
-#include "icm42688_spi.h"
+#include "icm4268x.h"
+#include "icm4268x_decoder.h"
+#include "icm4268x_reg.h"
+#include "icm4268x_rtio.h"
+#include "icm4268x_spi.h"
 
 #include <zephyr/logging/log.h>
-LOG_MODULE_REGISTER(ICM42688_RTIO, CONFIG_SENSOR_LOG_LEVEL);
+LOG_MODULE_REGISTER(ICM4268X_RTIO, CONFIG_SENSOR_LOG_LEVEL);
 
-static int icm42688_rtio_sample_fetch(const struct device *dev, int16_t readings[7])
+static int icm4268x_rtio_sample_fetch(const struct device *dev, int16_t readings[7])
 {
 	uint8_t status;
-	const struct icm42688_dev_cfg *cfg = dev->config;
+	const struct icm4268x_dev_cfg *cfg = dev->config;
 	uint8_t *buffer = (uint8_t *)readings;
 
-	int res = icm42688_spi_read(&cfg->spi, REG_INT_STATUS, &status, 1);
+	int res = icm4268x_spi_read(&cfg->spi, REG_INT_STATUS, &status, 1);
 
 	if (res) {
 		return res;
@@ -32,7 +32,7 @@ static int icm42688_rtio_sample_fetch(const struct device *dev, int16_t readings
 		return -EBUSY;
 	}
 
-	res = icm42688_read_all(dev, buffer);
+	res = icm4268x_read_all(dev, buffer);
 
 	if (res) {
 		return res;
@@ -45,16 +45,16 @@ static int icm42688_rtio_sample_fetch(const struct device *dev, int16_t readings
 	return 0;
 }
 
-static void icm42688_submit_one_shot(const struct device *dev, struct rtio_iodev_sqe *iodev_sqe)
+static void icm4268x_submit_one_shot(const struct device *dev, struct rtio_iodev_sqe *iodev_sqe)
 {
 	const struct sensor_read_config *cfg = iodev_sqe->sqe.iodev->data;
 	const struct sensor_chan_spec *const channels = cfg->channels;
 	const size_t num_channels = cfg->count;
-	uint32_t min_buf_len = sizeof(struct icm42688_encoded_data);
+	uint32_t min_buf_len = sizeof(struct icm4268x_encoded_data);
 	int rc;
 	uint8_t *buf;
 	uint32_t buf_len;
-	struct icm42688_encoded_data *edata;
+	struct icm4268x_encoded_data *edata;
 
 	/* Get the buffer for the frame, it may be allocated dynamically by the rtio context */
 	rc = rtio_sqe_rx_buf(iodev_sqe, min_buf_len, min_buf_len, &buf, &buf_len);
@@ -64,11 +64,11 @@ static void icm42688_submit_one_shot(const struct device *dev, struct rtio_iodev
 		return;
 	}
 
-	edata = (struct icm42688_encoded_data *)buf;
+	edata = (struct icm4268x_encoded_data *)buf;
 
-	icm42688_encode(dev, channels, num_channels, buf);
+	icm4268x_encode(dev, channels, num_channels, buf);
 
-	rc = icm42688_rtio_sample_fetch(dev, edata->readings);
+	rc = icm4268x_rtio_sample_fetch(dev, edata->readings);
 	/* Check that the fetch succeeded */
 	if (rc != 0) {
 		LOG_ERR("Failed to fetch samples");
@@ -79,27 +79,27 @@ static void icm42688_submit_one_shot(const struct device *dev, struct rtio_iodev
 	rtio_iodev_sqe_ok(iodev_sqe, 0);
 }
 
-void icm42688_submit_sync(struct rtio_iodev_sqe *iodev_sqe)
+void icm4268x_submit_sync(struct rtio_iodev_sqe *iodev_sqe)
 {
 	const struct sensor_read_config *cfg = iodev_sqe->sqe.iodev->data;
 	const struct device *dev = cfg->sensor;
 
 	if (!cfg->is_streaming) {
-		icm42688_submit_one_shot(dev, iodev_sqe);
-	} else if (IS_ENABLED(CONFIG_ICM42688_STREAM)) {
-		icm42688_submit_stream(dev, iodev_sqe);
+		icm4268x_submit_one_shot(dev, iodev_sqe);
+	} else if (IS_ENABLED(CONFIG_ICM4268X_STREAM)) {
+		icm4268x_submit_stream(dev, iodev_sqe);
 	} else {
 		rtio_iodev_sqe_err(iodev_sqe, -ENOTSUP);
 	}
 }
 
-void icm42688_submit(const struct device *dev, struct rtio_iodev_sqe *iodev_sqe)
+void icm4268x_submit(const struct device *dev, struct rtio_iodev_sqe *iodev_sqe)
 {
 	struct rtio_work_req *req = rtio_work_req_alloc();
 
 	__ASSERT_NO_MSG(req);
 
-	rtio_work_req_submit(req, iodev_sqe, icm42688_submit_sync);
+	rtio_work_req_submit(req, iodev_sqe, icm4268x_submit_sync);
 }
 
-BUILD_ASSERT(sizeof(struct icm42688_decoder_header) == 9);
+BUILD_ASSERT(sizeof(struct icm4268x_decoder_header) == 9);
