+++ drivers/sensor/tdk/icm4268x/icm4268x_rtio_stream.c
@@ -1,23 +1,24 @@
 /*
  * Copyright (c) 2023 Google LLC
+ * Copyright 2024 NXP
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
 #include <zephyr/logging/log.h>
 
-#include "icm42688.h"
-#include "icm42688_decoder.h"
-#include "icm42688_reg.h"
-#include "icm42688_rtio.h"
+#include "icm4268x.h"
+#include "icm4268x_decoder.h"
+#include "icm4268x_reg.h"
+#include "icm4268x_rtio.h"
 
-LOG_MODULE_DECLARE(ICM42688_RTIO);
+LOG_MODULE_DECLARE(ICM4268X_RTIO);
 
-void icm42688_submit_stream(const struct device *sensor, struct rtio_iodev_sqe *iodev_sqe)
+void icm4268x_submit_stream(const struct device *sensor, struct rtio_iodev_sqe *iodev_sqe)
 {
 	const struct sensor_read_config *cfg = iodev_sqe->sqe.iodev->data;
-	struct icm42688_dev_data *data = sensor->data;
-	struct icm42688_cfg new_config = data->cfg;
+	struct icm4268x_dev_data *data = sensor->data;
+	struct icm4268x_cfg new_config = data->cfg;
 
 	new_config.interrupt1_drdy = false;
 	new_config.interrupt1_fifo_ths = false;
@@ -43,7 +44,7 @@ void icm42688_submit_stream(const struct device *sensor, struct rtio_iodev_sqe *
 	if (new_config.interrupt1_drdy != data->cfg.interrupt1_drdy ||
 	    new_config.interrupt1_fifo_ths != data->cfg.interrupt1_fifo_ths ||
 	    new_config.interrupt1_fifo_full != data->cfg.interrupt1_fifo_full) {
-		int rc = icm42688_safely_configure(sensor, &new_config);
+		int rc = icm4268x_safely_configure(sensor, &new_config);
 
 		if (rc != 0) {
 			LOG_ERR("Failed to configure sensor");
@@ -55,11 +56,11 @@ void icm42688_submit_stream(const struct device *sensor, struct rtio_iodev_sqe *
 	data->streaming_sqe = iodev_sqe;
 }
 
-static void icm42688_complete_cb(struct rtio *r, const struct rtio_sqe *sqe, void *arg)
+static void icm4268x_complete_cb(struct rtio *r, const struct rtio_sqe *sqe, void *arg)
 {
 	const struct device *dev = arg;
-	struct icm42688_dev_data *drv_data = dev->data;
-	const struct icm42688_dev_cfg *drv_cfg = dev->config;
+	struct icm4268x_dev_data *drv_data = dev->data;
+	const struct icm4268x_dev_cfg *drv_cfg = dev->config;
 	struct rtio_iodev_sqe *iodev_sqe = sqe->userdata;
 
 	rtio_iodev_sqe_ok(iodev_sqe, drv_data->fifo_count);
@@ -67,11 +68,11 @@ static void icm42688_complete_cb(struct rtio *r, const struct rtio_sqe *sqe, voi
 	gpio_pin_interrupt_configure_dt(&drv_cfg->gpio_int1, GPIO_INT_EDGE_TO_ACTIVE);
 }
 
-static void icm42688_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, void *arg)
+static void icm4268x_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, void *arg)
 {
 	const struct device *dev = arg;
-	struct icm42688_dev_data *drv_data = dev->data;
-	const struct icm42688_dev_cfg *drv_cfg = dev->config;
+	struct icm4268x_dev_data *drv_data = dev->data;
+	const struct icm4268x_dev_cfg *drv_cfg = dev->config;
 	struct rtio_iodev *spi_iodev = drv_data->spi_iodev;
 	uint8_t *fifo_count_buf = (uint8_t *)&drv_data->fifo_count;
 	uint16_t fifo_count = ((fifo_count_buf[0] << 8) | fifo_count_buf[1]);
@@ -91,8 +92,8 @@ static void icm42688_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, v
 	}
 
 	const size_t packet_size = drv_data->cfg.fifo_hires ? 20 : 16;
-	const size_t min_read_size = sizeof(struct icm42688_fifo_data) + packet_size;
-	const size_t ideal_read_size = sizeof(struct icm42688_fifo_data) + fifo_count;
+	const size_t min_read_size = sizeof(struct icm4268x_fifo_data) + packet_size;
+	const size_t ideal_read_size = sizeof(struct icm4268x_fifo_data) + fifo_count;
 	uint8_t *buf;
 	uint32_t buf_len;
 
@@ -108,7 +109,7 @@ static void icm42688_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, v
 	/* TODO is packet format even needed? the fifo has a header per packet
 	 * already
 	 */
-	struct icm42688_fifo_data hdr = {
+	struct icm4268x_fifo_data hdr = {
 		.header = {
 			.is_fifo = true,
 			.gyro_fs = drv_data->cfg.gyro_fs,
@@ -128,7 +129,7 @@ static void icm42688_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, v
 	uint32_t pkts = read_len / packet_size;
 
 	read_len = pkts * packet_size;
-	((struct icm42688_fifo_data *)buf)->fifo_count = read_len;
+	((struct icm4268x_fifo_data *)buf)->fifo_count = read_len;
 
 	__ASSERT_NO_MSG(read_len % pkt_size == 0);
 
@@ -157,13 +158,13 @@ static void icm42688_fifo_count_cb(struct rtio *r, const struct rtio_sqe *sqe, v
 	rtio_sqe_prep_read(read_fifo_data, spi_iodev, RTIO_PRIO_NORM, read_buf, read_len,
 			   iodev_sqe);
 	read_fifo_data->flags = RTIO_SQE_CHAINED;
-	rtio_sqe_prep_callback(complete_op, icm42688_complete_cb, (void *)dev, iodev_sqe);
+	rtio_sqe_prep_callback(complete_op, icm4268x_complete_cb, (void *)dev, iodev_sqe);
 
 	rtio_submit(r, 0);
 }
 
 static struct sensor_stream_trigger *
-icm42688_get_read_config_trigger(const struct sensor_read_config *cfg,
+icm4268x_get_read_config_trigger(const struct sensor_read_config *cfg,
 				 enum sensor_trigger_type trig)
 {
 	for (int i = 0; i < cfg->count; ++i) {
@@ -175,11 +176,11 @@ icm42688_get_read_config_trigger(const struct sensor_read_config *cfg,
 	return NULL;
 }
 
-static void icm42688_int_status_cb(struct rtio *r, const struct rtio_sqe *sqr, void *arg)
+static void icm4268x_int_status_cb(struct rtio *r, const struct rtio_sqe *sqr, void *arg)
 {
 	const struct device *dev = arg;
-	struct icm42688_dev_data *drv_data = dev->data;
-	const struct icm42688_dev_cfg *drv_cfg = dev->config;
+	struct icm4268x_dev_data *drv_data = dev->data;
+	const struct icm4268x_dev_cfg *drv_cfg = dev->config;
 	struct rtio_iodev *spi_iodev = drv_data->spi_iodev;
 	struct rtio_iodev_sqe *streaming_sqe = drv_data->streaming_sqe;
 	struct sensor_read_config *read_config;
@@ -197,12 +198,12 @@ static void icm42688_int_status_cb(struct rtio *r, const struct rtio_sqe *sqr, v
 	}
 
 	struct sensor_stream_trigger *fifo_ths_cfg =
-		icm42688_get_read_config_trigger(read_config, SENSOR_TRIG_FIFO_WATERMARK);
+		icm4268x_get_read_config_trigger(read_config, SENSOR_TRIG_FIFO_WATERMARK);
 	bool has_fifo_ths_trig = fifo_ths_cfg != NULL &&
 				 FIELD_GET(BIT_INT_STATUS_FIFO_THS, drv_data->int_status) != 0;
 
 	struct sensor_stream_trigger *fifo_full_cfg =
-		icm42688_get_read_config_trigger(read_config, SENSOR_TRIG_FIFO_FULL);
+		icm4268x_get_read_config_trigger(read_config, SENSOR_TRIG_FIFO_FULL);
 	bool has_fifo_full_trig = fifo_full_cfg != NULL &&
 				  FIELD_GET(BIT_INT_STATUS_FIFO_FULL, drv_data->int_status) != 0;
 
@@ -240,13 +241,13 @@ static void icm42688_int_status_cb(struct rtio *r, const struct rtio_sqe *sqr, v
 
 		/* Clear streaming_sqe since we're done with the call */
 		drv_data->streaming_sqe = NULL;
-		if (rtio_sqe_rx_buf(streaming_sqe, sizeof(struct icm42688_fifo_data),
-				    sizeof(struct icm42688_fifo_data), &buf, &buf_len) != 0) {
+		if (rtio_sqe_rx_buf(streaming_sqe, sizeof(struct icm4268x_fifo_data),
+				    sizeof(struct icm4268x_fifo_data), &buf, &buf_len) != 0) {
 			rtio_iodev_sqe_err(streaming_sqe, -ENOMEM);
 			return;
 		}
 
-		struct icm42688_fifo_data *data = (struct icm42688_fifo_data *)buf;
+		struct icm4268x_fifo_data *data = (struct icm4268x_fifo_data *)buf;
 
 		memset(buf, 0, buf_len);
 		data->header.timestamp = drv_data->timestamp;
@@ -282,14 +283,14 @@ static void icm42688_int_status_cb(struct rtio *r, const struct rtio_sqe *sqr, v
 	write_fifo_count_reg->flags = RTIO_SQE_TRANSACTION;
 	rtio_sqe_prep_read(read_fifo_count, spi_iodev, RTIO_PRIO_NORM, read_buf, 2, NULL);
 	read_fifo_count->flags = RTIO_SQE_CHAINED;
-	rtio_sqe_prep_callback(check_fifo_count, icm42688_fifo_count_cb, arg, NULL);
+	rtio_sqe_prep_callback(check_fifo_count, icm4268x_fifo_count_cb, arg, NULL);
 
 	rtio_submit(r, 0);
 }
 
-void icm42688_fifo_event(const struct device *dev)
+void icm4268x_fifo_event(const struct device *dev)
 {
-	struct icm42688_dev_data *drv_data = dev->data;
+	struct icm4268x_dev_data *drv_data = dev->data;
 	struct rtio_iodev *spi_iodev = drv_data->spi_iodev;
 	struct rtio *r = drv_data->r;
 
@@ -317,6 +318,6 @@ void icm42688_fifo_event(const struct device *dev)
 	write_int_reg->flags = RTIO_SQE_TRANSACTION;
 	rtio_sqe_prep_read(read_int_reg, spi_iodev, RTIO_PRIO_NORM, &drv_data->int_status, 1, NULL);
 	read_int_reg->flags = RTIO_SQE_CHAINED;
-	rtio_sqe_prep_callback(check_int_status, icm42688_int_status_cb, (void *)dev, NULL);
+	rtio_sqe_prep_callback(check_int_status, icm4268x_int_status_cb, (void *)dev, NULL);
 	rtio_submit(r, 0);
 }
