+++ drivers/sensor/tdk/icm4268x/icm4268x_spi.c
@@ -1,12 +1,13 @@
 /*
  * Copyright (c) 2022 Intel Corporation
+ * Copyright 2024 NXP
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
 #include <zephyr/sys/util.h>
-#include "icm42688_spi.h"
-#include "icm42688_reg.h"
+#include "icm4268x_spi.h"
+#include "icm4268x_reg.h"
 
 static inline int spi_write_register(const struct spi_dt_spec *bus, uint8_t reg, uint8_t data)
 {
@@ -63,7 +64,7 @@ static inline int spi_read_register(const struct spi_dt_spec *bus, uint8_t reg, 
 	return spi_transceive_dt(bus, &tx, &rx);
 }
 
-int icm42688_spi_read(const struct spi_dt_spec *bus, uint16_t reg, uint8_t *data, size_t len)
+int icm4268x_spi_read(const struct spi_dt_spec *bus, uint16_t reg, uint8_t *data, size_t len)
 {
 	int res = 0;
 	uint8_t address = FIELD_GET(REG_ADDRESS_MASK, reg);
@@ -73,11 +74,11 @@ int icm42688_spi_read(const struct spi_dt_spec *bus, uint16_t reg, uint8_t *data
 	return res;
 }
 
-int icm42688_spi_update_register(const struct spi_dt_spec *bus, uint16_t reg, uint8_t mask,
+int icm4268x_spi_update_register(const struct spi_dt_spec *bus, uint16_t reg, uint8_t mask,
 				 uint8_t data)
 {
 	uint8_t temp = 0;
-	int res = icm42688_spi_read(bus, reg, &temp, 1);
+	int res = icm4268x_spi_read(bus, reg, &temp, 1);
 
 	if (res) {
 		return res;
@@ -86,10 +87,10 @@ int icm42688_spi_update_register(const struct spi_dt_spec *bus, uint16_t reg, ui
 	temp &= ~mask;
 	temp |= FIELD_PREP(mask, data);
 
-	return icm42688_spi_single_write(bus, reg, temp);
+	return icm4268x_spi_single_write(bus, reg, temp);
 }
 
-int icm42688_spi_single_write(const struct spi_dt_spec *bus, uint16_t reg, uint8_t data)
+int icm4268x_spi_single_write(const struct spi_dt_spec *bus, uint16_t reg, uint8_t data)
 {
 	int res = 0;
 	uint8_t address = FIELD_GET(REG_ADDRESS_MASK, reg);
