+++ tests/drivers/build_all/sensor/sensors_trigger_global.conf
@@ -25,7 +25,7 @@ CONFIG_HMC5883L_TRIGGER_GLOBAL_THREAD=y
 CONFIG_HTS221_TRIGGER_GLOBAL_THREAD=y
 CONFIG_ICM42605_TRIGGER_GLOBAL_THREAD=y
 CONFIG_ICM42670_TRIGGER_GLOBAL_THREAD=y
-CONFIG_ICM42688_TRIGGER_GLOBAL_THREAD=y
+CONFIG_ICM4268X_TRIGGER_GLOBAL_THREAD=y
 CONFIG_IIS2DH_TRIGGER_GLOBAL_THREAD=y
 CONFIG_IIS2DLPC_TRIGGER_GLOBAL_THREAD=y
 CONFIG_IIS2ICLX_TRIGGER_GLOBAL_THREAD=y
