+++ tests/drivers/build_all/sensor/sensors_trigger_none.conf
@@ -25,7 +25,7 @@ CONFIG_HMC5883L_TRIGGER_NONE=y
 CONFIG_HTS221_TRIGGER_NONE=y
 CONFIG_ICM42605_TRIGGER_NONE=y
 CONFIG_ICM42670_TRIGGER_NONE=y
-CONFIG_ICM42688_TRIGGER_NONE=y
+CONFIG_ICM4268X_TRIGGER_NONE=y
 CONFIG_IIS2DH_TRIGGER_NONE=y
 CONFIG_IIS2DLPC_TRIGGER_NONE=y
 CONFIG_IIS2ICLX_TRIGGER_NONE=y
