+++ tests/drivers/build_all/sensor/sensors_trigger_own.conf
@@ -23,7 +23,7 @@ CONFIG_GROW_R502A_TRIGGER_OWN_THREAD=y
 CONFIG_HMC5883L_TRIGGER_OWN_THREAD=y
 CONFIG_HTS221_TRIGGER_OWN_THREAD=y
 CONFIG_ICM42670_TRIGGER_OWN_THREAD=y
-CONFIG_ICM42688_TRIGGER_OWN_THREAD=y
+CONFIG_ICM4268X_TRIGGER_OWN_THREAD=y
 CONFIG_IIS2DH_TRIGGER_OWN_THREAD=y
 CONFIG_IIS2DLPC_TRIGGER_OWN_THREAD=y
 CONFIG_IIS2ICLX_TRIGGER_OWN_THREAD=y
