+++ tests/drivers/sensor/icm4268x/boards/native_sim.overlay
@@ -3,9 +3,16 @@
  */
 
 &spi0 {
+	icm42686: icm42686@2 {
+		compatible = "invensense,icm42686";
+		int-gpios = <&gpio0 1 GPIO_ACTIVE_HIGH>;
+		spi-max-frequency = <50000000>;
+		reg = <2>;
+	};
+
 	icm42688: icm42688@3 {
 		compatible = "invensense,icm42688";
-		int-gpios = <&gpio0 1 GPIO_ACTIVE_HIGH>;
+		int-gpios = <&gpio0 2 GPIO_ACTIVE_HIGH>;
 		spi-max-frequency = <50000000>;
 		reg = <3>;
 	};
