+++ tests/drivers/sensor/icm4268x/prj.conf
